Add CSS animations
An animation called "fade-in" has been added to the h2 heading. The animation gradually makes the initially invisible heading become visible. To use the animation, the h2 element has been given the style rule "animation: fade-in 1s ease". This way, the h2 heading will become visible with the animation when the page is loaded or displayed.